LPS-188073 Adapt code to the last master changes
