+++ modules/apps/headless/headless-builder/headless-builder-test/src/testIntegration/java/com/liferay/headless/builder/application/resource/test/HeadlessBuilderResourceTest.java
@@ -12,12 +12,12 @@
  * details.
  */
 
-package com.liferay.headless.builder.resource.test;
+package com.liferay.headless.builder.application.resource.test;
 
 import com.liferay.headless.builder.application.APIApplication;
 import com.liferay.headless.builder.application.provider.APIApplicationProvider;
 import com.liferay.headless.builder.application.publisher.APIApplicationPublisher;
-import com.liferay.headless.builder.publisher.test.util.APIApplicationPublisherUtil;
+import com.liferay.headless.builder.application.publisher.test.util.APIApplicationPublisherUtil;
 import com.liferay.headless.builder.test.BaseTestCase;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.json.JSONUtil;
@@ -58,18 +58,6 @@ public class HeadlessBuilderResourceTest extends BaseTestCase {
 
 	@After
 	public void tearDown() throws Exception {
-		HTTPTestUtil.invoke(
-			null,
-			"headless-builder/applications/by-external-reference-code/" +
-				_API_APPLICATION_ERC_1,
-			Http.Method.DELETE);
-
-		HTTPTestUtil.invoke(
-			null,
-			"headless-builder/applications/by-external-reference-code/" +
-				_API_APPLICATION_ERC_2,
-			Http.Method.DELETE);
-
 		APIApplicationPublisherUtil.unpublishRemainingAPIApplications(
 			_apiApplicationPublisher);
 	}
